Create SearchInput Component and update admin/index.tsx, packages.tsx, and products.tsx
